Refactor tests into appropriate package
